Move DataViewBasicInfoPanel to components/index/cards/charts
