+++ components/index/cards/charts/ConfirmedCasesByMunicipalitiesTable.vue
@@ -69,7 +69,7 @@
 <script lang="ts">
 import Vue from 'vue'
 import DataView from '@/components/DataView.vue'
-import DataViewBasicInfoPanel from '@/components/DataViewBasicInfoPanel.vue'
+import DataViewBasicInfoPanel from '@/components/index/cards/charts/DataViewBasicInfoPanel.vue'
 
 export default Vue.extend({
   components: { DataView, DataViewBasicInfoPanel },
