+++ components/index/cards/charts/DataViewBasicInfoPanel.vue
